Zutaten anlegen - check

+++ src/main/java/com/meinname/kochbuch/model/Zutat.java
@@ -7,7 +7,7 @@ public class Zutat {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String name;
@@ -16,18 +16,15 @@ public class Zutat {
     public Zutat() {
     }
 
-    // Konstruktor ohne ID
     public Zutat(String name) {
         this.name = name;
     }
 
-    // Getter und Setter
-
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
